+++ admin/views/sms-templates/edit-template.php
@@ -39,10 +39,12 @@ $repVarArr = !empty($fld->value) && LibHelper::isJson($fld->value) ? json_decode
 $repVarArr = is_array($repVarArr) ? $repVarArr : [];
 if (!empty($repVarArr)) {
     $repVarHtml = '<ul class="click-to-copy">';
+    $found = false;
     foreach ($repVarArr as $val) {
         if (!isset($val['variable'])) {
             continue;
         }
+        $found = true;
         $repVarHtml .= '<li title="' . Labels::getLabel('LBL_CLICK_TO_COPY', $siteLangId) . '" onclick="copyText(this, true);" data-bs-toggle="tooltip" data-placement="top">
             <div class="text">
                 <span>' . $val['title'] . '</span>
@@ -51,6 +53,9 @@ if (!empty($repVarArr)) {
                 </span>
             </div>
         </li>';
+    }
+    if (false == $found) {
+        $repVarHtml .= '<li>' . Labels::getLabel('LBL_N/A') . '</li>';
     }
     $repVarHtml .= '</ul>';
     $fld->value = $repVarHtml;