+++ CHANGELOG.md
@@ -38,6 +38,9 @@
     - Bug-082754 - When the 'Available Quantity' is 0, the buyer is unable to place the RFQ order
     - Bug-082765 - No validation message appears when buyer does not add Delivery Address on 'Request a quote' popup box.
     - Bug-082762 - The unit of the ordered quantity should be displayed on the 'Your Shopping Bag' page of buyer
-
+    - Bug-080244 - Error occur on the My Offers page on the buyer dashboard in Arabic language.
+    - Bug-079388 - "Pay At Store" not working.
 ## Hit URL:
-    - admin/patch-update/updateProductRating    
+    - admin/patch-update/updateProductRating  
+    
+